pyproject.toml
[project]
name = "claude-company"
version = "0.1.0"
description = "Claude Company - note.com article tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "youtube-transcript-api>=1.2.4",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.44.0",
    "python-frontmatter>=1.1.0",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["note_preview"]

[project.scripts]
note-preview = "note_preview.cli:main"
